chore: BasicAuthorizationExtractor 로 인터셉터와 아규먼트 리졸버의 중복 제거
