+++ src/main/java/cart/controller/CartController.java
@@ -1,8 +1,8 @@
 package cart.controller;
 
-import cart.controller.argumentresolver.BasicAuthorization;
+import cart.controller.auth.argumentresolver.BasicAuthorization;
+import cart.controller.auth.dto.LoginUser;
 import cart.controller.dto.CartResponse;
-import cart.controller.dto.LoginUser;
 import cart.service.CartService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
