+++ src/test/java/cart/controller/CartControllerTest.java
@@ -1,7 +1,8 @@
 package cart.controller;
 
+import cart.controller.auth.dto.LoginUser;
+import cart.controller.auth.dto.LoginUser;
 import cart.controller.dto.CartResponse;
-import cart.controller.dto.LoginUser;
 import cart.domain.*;
 import cart.service.CartService;
 import com.fasterxml.jackson.databind.ObjectMapper;
